Renamed 'TradingBot.TechnicalAnalyze.Core' and 'TradingBot.TechnicalAnalyze.Tests' projects to 'TradingBot.TechnicalAnalyze' and 'TradingBot.TechnicalAnalyze.Tests'. Created 'TradingBot.Simulator' project.
