+++ TradingBot.Infrastructure/TradingBot.Infrastructure.csproj
@@ -9,7 +9,7 @@
   <ItemGroup>
     <ProjectReference Include="..\TradingBot.Core\TradingBot.Core.csproj" />
     <ProjectReference Include="..\TradingBot.HttpClients\TradingBot.HttpClients.csproj" />
-    <ProjectReference Include="..\TradingBot.TechnicalAnalyze.Core\TradingBot.TechnicalAnalyze.Core.csproj" />
+    <ProjectReference Include="..\TradingBot.TechnicalAnalyze\TradingBot.TechnicalAnalyze.csproj" />
   </ItemGroup>
 
   <ItemGroup>
