+++ TradingBot.TechnicalAnalyze.Tests/TradingBot.TechnicalAnalyze.Tests.csproj
@@ -25,7 +25,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\TradingBot.HttpClients\TradingBot.HttpClients.csproj" />
-    <ProjectReference Include="..\TradingBot.TechnicalAnalyze.Core\TradingBot.TechnicalAnalyze.Core.csproj" />
+    <ProjectReference Include="..\TradingBot.TechnicalAnalyze\TradingBot.TechnicalAnalyze.csproj" />
   </ItemGroup>
 
 </Project>
