+++ TradingBot.TechnicalAnalyze/TradingBot.TechnicalAnalyze.csproj
